requirements/requirements.txt
aliyun-python-sdk-bssopenapi==1.2.8
aliyun-python-sdk-core-v3==2.13.3
aliyun-python-sdk-ecs==4.10.1
aliyun-python-sdk-r-kvstore==2.1.1
aliyun-python-sdk-rds==2.3.9
aliyun-python-sdk-slb==3.2.10
amqp==2.1.4
ansible==2.8.0
asn1crypto==0.24.0
azure-common==1.1.22
azure-nspkg==3.0.2
azure-storage-blob==1.3.1
azure-storage-common==1.4.2
azure-storage-nspkg==3.1.0
bcrypt==3.1.4
billiard==3.5.0.3
boto==2.49.0
boto3==1.6.5
botocore==1.9.5
celery==4.1.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
configparser==3.5.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==2.3.1
decorator==4.1.2
Django==2.1.7
django-auth-ldap==1.7.0
django-bootstrap3==9.1.0
django-celery-beat==1.4.0
django-filter==2.0.0
django-formtools==2.1
django-radius==1.3.3
django-ranged-response==0.2.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-swagger==2.1.2
django-simple-captcha==0.5.6
django-timezone-field==3.0
djangorestframework==3.9.4
djangorestframework-bulk==0.2.1
dnspython==1.16.0
docutils==0.14
drf-nested-routers==0.91
drf-yasg==1.9.1
ecdsa==0.13
elasticsearch==6.1.1
enum-compat==0.0.2
ephem==3.7.6.0
eventlet==0.24.1
ForgeryPy==0.1
future==0.17.1
greenlet==0.4.14
gunicorn==19.9.0
idna==2.6
inflection==0.3.1
ipip-ipdb==1.2.1
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10.1
jmespath==0.9.3
jms-storage==0.0.23
kombu==4.0.2
ldap3==2.4
MarkupSafe==1.0
monotonic==1.5
mysqlclient==1.3.14
netaddr==0.7.19
olefile==0.44
openapi-codec==1.3.2
oss2==2.4.0
paramiko==2.4.2
passlib==1.7.1
Pillow==4.3.0
pyasn1==0.4.2
pyasn1-modules==0.2.5
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.8.2
PyNaCl==1.2.1
pyotp==2.2.6
pyrad==2.1
python-crontab==2.3.6
python-dateutil==2.6.1
python-gssapi==0.6.4
python-jose==3.0.1
python-keycloak==0.13.3
python-keycloak-client==0.1.3
python-ldap==3.1.0
pytz==2018.3
PyYAML==5.1
redis==2.10.6
requests==2.22.0
rest-condition==1.0.3
rsa==4.0
ruamel.yaml==0.15.97
s3transfer==0.1.13
simplejson==3.13.2
six==1.11.0
sshpubkeys==3.1.0
tencentcloud-sdk-python==3.0.40
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.2
vine==1.1.4
Werkzeug==0.14.1
